requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
aiohttp-cors==0.7.0
ajpy==0.0.4
appdirs==1.4.3
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
backdoor-factory==0.0.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
binwalk==2.1.2
black==19.10b0
blinker==1.4
Brlapi==0.7.0
Brotli==1.0.7
capstone==4.0.1
certifi==2020.4.5.1
chardet==3.0.4
chatterbot-corpus==1.2.0
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crcelk==1.3
cryptography==2.3
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debtags==2.1
decorator==4.3.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.3.0
distro-info==0.21
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.42.0
django-dotenv==1.4.2
django-filter==2.2.0
django-heroku==0.3.1
django-sendgrid-v5==0.9.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
drf-yasg==1.17.1
entrypoints==0.3
Faker==4.1.0
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
future==0.16.0
gns3-gui==2.2.10
gns3-server==2.2.10
gunicorn==20.0.4
h11==0.8.1
h2==3.0.1
hashID==3.1.4
hpack==3.0.0
html5lib==1.0.1
httplib2==0.11.3
hyperframe==5.1.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==16.10.1
inflection==0.4.0
IPy==0.83
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-core==4.4.0
kaitaistruct==0.8
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
ldap3==2.4.1
louis==3.8.0
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mitmproxy==4.0.4
more-itertools==4.2.0
multidict==4.7.6
mysqlclient==1.4.4
nassl==3.0.0
nbformat==4.4.0
numpy==1.18.3
oauthlib==3.1.0
olefile==0.46
packaging==20.4
paramiko==2.4.2
passlib==1.7.1
patator==0.7
pathspec==0.8.0
pbr==5.4.5
pefile==2018.8.8
Pillow==7.0.0
plotly==3.6.1
pluginbase==1.0.0
ply==3.11
psutil==5.6.6
psycopg2==2.8.5
py-cpuinfo==5.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pycurl==7.43.0.2
pyenchant==2.0.0
PyGObject==3.36.0
pygtkspellcheck==4.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.5.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyperclip==1.6.4
PyQt5==5.15.0
pyrsistent==0.16.0
pyserial==3.4
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
pysmi==0.3.2
pysnmp==4.4.6
PySocks==1.6.8
python-apt==2.1.3
python-dateutil==2.8.1
python-debian==0.1.34
python-debianbts==2.8.2
python-decouple==3.3
python-http-client==3.3.1
python-pam==1.8.4
python3-openid==3.2.0
pytz==2020.1
pyxattr==0.6.1
pyxdg==0.25
PyYAML==3.13
regex==2020.6.8
reportbug==7.5.2
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scapy==2.4.3
scipy==1.4.1
scour==0.37
SecretStorage==2.3.1
sendgrid==6.4.6
sentry-sdk==0.15.1
service-identity==18.1.0
sip==4.19.22
six==1.15.0
smoke-zephyr==1.4.0
sortedcontainers==2.0.4
soupsieve==1.8
SQLAlchemy==1.3.18
sqlparse==0.3.1
sslyze==2.0.6
starkbank-ecdsa==1.0.0
stevedore==1.31.0
tabulate==0.8.2
termcolor==1.1.0
termineter==1.0.4
text-unidecode==1.3
tld==0.9.1
tls-parser==1.2.1
toml==0.10.1
tornado==5.1.1
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.4.1
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.9
urwid==2.0.1
vboxapi==1.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wafw00f==0.9.5
wapiti3==3.0.1
webencodings==0.5.1
Werkzeug==1.0.1
wfuzz==2.3.4
whitenoise==5.2.0
wrapt==1.12.1
wsproto==0.11.0
yarl==1.4.2
yaswfp==0.9.3
youtube-dl==2019.1.17
zipp==1.0.0
zope.interface==4.7.1
